feat(sim): plants at real scale (Uint16 height, cropped at canvas top)
Switch plantHeight to Uint16Array and calibrate PLANT_MAX_HEIGHT to
real *P. barbatus* habitat vegetation at 3 mm/cell:

  - grass clumps: ~30 cm = 100 cells
  - shrubs (creosote, brittlebush): ~1.2 m = 400 cells
  - trees (mesquite, palo verde): ~4.5 m = 1500 cells

These exceed the visible above-surface band; the GL fragment shader
already crops naturally via the sampling bounds (cell.y always in
[0, h-1] regardless of how negative plantTop goes).

Trunk proportions adjusted for the new scale so the cropped view
reads correctly:
  - grass: 1-cell stem at base, rest canopy → tufts on the surface
  - shrub: 2-cell woody base, rest canopy → bushy domes
  - tree: ~h/4 trunk → brown column rising off-screen, no canopy
    visible (matches an ant looking up at a mesquite trunk)

uPlantHeight is now usampler2D R16UI, mirroring uSurf. Persist
schema bumps v10 → v11 (Uint16 byte layout).

(In-flight; speed-cap bump deferred to a follow-up.)

+++ src/main.ts
@@ -58,7 +58,7 @@ function makePlaceholderWorld(s: Settings): {
   corpse: Uint8Array; sprout: Uint8Array; sproutTick: Int32Array;
   digTick: Int32Array;
   plant: Uint8Array;
-  plantHeight: Uint8Array;
+  plantHeight: Uint16Array;
 } {
   // Renderer needs SOMETHING at construction time, before the
   // worker has produced its first snapshot. Empty arrays of the
@@ -75,7 +75,7 @@ function makePlaceholderWorld(s: Settings): {
     sprout: new Uint8Array(sz), sproutTick: new Int32Array(sz),
     digTick: new Int32Array(sz),
     plant: new Uint8Array(s.width),
-    plantHeight: new Uint8Array(s.width),
+    plantHeight: new Uint16Array(s.width),
   };
 }
